simplify space tests to prevent flakiness
