+++ x-pack/test/fleet_api_integration/apis/epm/index.js
@@ -19,7 +19,7 @@ export default function loadTests({ loadTestFile }) {
     loadTestFile(require.resolve('./install_overrides'));
     loadTestFile(require.resolve('./install_prerelease'));
     loadTestFile(require.resolve('./install_remove_assets'));
-    loadTestFile(require.resolve('./install_remove_assets_in_space'));
+    loadTestFile(require.resolve('./install_remove_kbn_assets_in_space'));
     loadTestFile(require.resolve('./install_remove_multiple'));
     loadTestFile(require.resolve('./install_update'));
     loadTestFile(require.resolve('./bulk_upgrade'));
