+++ x-pack/test/fleet_api_integration/apis/fixtures/test_packages/only_dashboard/0.1.0/docs/README.md
@@ -1,0 +1,3 @@
+# Test package
+
+For testing that a package installs its elasticsearch assets when installed for the first time (not updating) and removing the package
