+++ x-pack/test/fleet_api_integration/apis/fixtures/test_packages/only_dashboard/0.1.0/manifest.yml
@@ -1,0 +1,20 @@
+format_version: 1.0.0
+name: only_dashboard 
+title: Only installs one dashboard
+description: This package is used to test installing assets into a non-default space.
+version: 0.1.0
+categories: []
+release: beta
+type: integration
+license: basic
+
+requirement:
+  elasticsearch:
+    versions: '>7.7.0'
+  kibana:
+    versions: '>7.7.0'
+
+icons:
+  - src: '/img/logo_overrides_64_color.svg'
+    size: '16x16'
+    type: 'image/svg+xml'
